Trait tower_h2::Body

pub trait Body {
    type Data: IntoBuf + 'static;
    fn poll_data(&mut self) -> Poll<Option<Self::Data>, Error>;

    fn is_end_stream(&self) -> bool { ... }

    fn poll_trailers(&mut self) -> Poll<Option<HeaderMap>, Error> { ... }
}

A generic h2 client/server request/response body.

Associated Types

type Data: IntoBuf + 'static

The body chunk type

Required Methods

fn poll_data(&mut self) -> Poll<Option<Self::Data>, Error>

Polls a stream of data.

Provided Methods

fn is_end_stream(&self) -> bool

Returns true when the end of stream has been reached.

An end of stream means that both poll_data and poll_trailers will return None.

A return value of false does not guarantee that a value will be returend from poll_stream or poll_trailers.

fn poll_trailers(&mut self) -> Poll<Option<HeaderMap>, Error>

Returns possibly one HeaderMap for trailers.