Struct tower_h2::client::Connect

pub struct Connect<C, E, S> { /* fields omitted */ }

Establishes an H2 client connection.

Has a builder-like API for configuring client connections. Currently this only allows the configuration of TLS transport on new services created by this factory.

Methods

impl<C, E, S> Connect<C, E, S> where
C: Connect,
E: Executor<Background<C::Connected, S>> + Clone,
S: Body,

pub fn new(inner: C, builder: Builder, executor: E) -> Self

Create a new Connect.

The connect argument is used to obtain new session layer instances (AsyncRead + AsyncWrite). For each new client service returned, a task will be spawned onto executor that will be used to manage the H2 connection.