Trait tower_h2::HttpService

pub trait HttpService: Sealed {
    type RequestBody: Body;
    type ResponseBody: Body;
    type Error;
    type Future: Future<Item = Response<Self::ResponseBody>, Error = Self::Error>;
    fn poll_ready(&mut self) -> Poll<(), Self::Error>;

    fn call(&mut self, request: Request<Self::RequestBody>) -> Self::Future;

    fn lift(self) -> LiftService<Self>
    where
        Self: Sized,
    { ... }

    fn lift_ref(&mut self) -> LiftServiceRef<Self>
    where
        Self: Sized,
    { ... }
}

An HTTP service

This is not intended to be implemented directly. Instead, it is a trait alias of sorts. Implements the tower_service::Service trait using http::Request and http::Response types.

Associated Types

type RequestBody: Body

Request payload.

type ResponseBody: Body

Response payload.

type Error

Errors produced by the service.

type Future: Future<Item = Response<Self::ResponseBody>, Error = Self::Error>

The future response value.

Required Methods

fn poll_ready(&mut self) -> Poll<(), Self::Error>

Returns Ready when the service is able to process requests.

fn call(&mut self, request: Request<Self::RequestBody>) -> Self::Future

Process the request and return the response asynchronously.

Provided Methods

fn lift(self) -> LiftService<Self> where
Self: Sized,

Wrap the HttpService so that it implements tower_service::Service directly.

Since HttpService does not directly implement Service, if an HttpService instance needs to be used where a T: Service is required, it must be wrapped with a type that provides that implementation. LiftService does this.

fn lift_ref(&mut self) -> LiftServiceRef<Self> where
Self: Sized,

Same as lift but operates on an HttpService reference.